requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary
python-decouple==3.8
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
setuptools